all skeleton tested

+++ src/main/java/org/wizindia/black/validation/FileSizeValidator.java
@@ -17,7 +17,7 @@ public class FileSizeValidator implements Validator {
 
     @Override
     public List<? extends ValidationError> validate(Object object) {
-        int size = (int)object;
+        long size = (long)object;
         if (size < fileMinSize)
             return Arrays.asList(ValidationErrorCode.FileSizeLessThanExpected);
         if (size > fileMaxSize)
